+++ extensions/raycast-ollama/CHANGELOG.md
@@ -1,3 +1,10 @@
 # raycast-ollama Changelog
 
 ## [Initial Version] - 2023-07-28
+
+## [Code Improvement and BugFix] - 2023-07-29
+
+- [BugFix] Fixed custom model file path and base model name for llama2.
+- [Code Improvement] Better error handling.
+- [Code Improvement] Added JSDoc documentation.
+- [Code Improvement] Errors, Types and Ollama functions moved out from main.tsx file.